+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.9.12
 * Add .perDevice queue
 * Improve code level documentation
+* Rename ScanResult to BleDevice
+* Add isConnected property to BleDevice
 
 ## 0.9.11
 * Add device name prefix filtering